Update i18n-routing.md - NextJS v12 middleware redirect to infinite loop error
In the section **Prefixing the Default Locale** for the use of localization with the Next v12 middleware, there is an error where if the redirect happens the url becomes an infinite loop of that url. 
That is because it says redirect to **NextResponse.redirect(`/en${request.nextUrl.href}`)** however request.nextUrl.href returns the whole url instead of just the path. I think it should say **request.nextUrl.pathname** which returns only the path.

I saw some issues about this but couldn't find if it had been resolved. Anyways, hope it helps 👍